Remove unused getDefaultParam method from HttpCommand and use TS for tests
The method was not used at all and therefore could be removed. Update
the HttpCommand tests to use TypeScript and createHttp testing function.
